requirements.txt
pandas 
faiss-cpu 
sentence-transformers 
openai 
geopy 
python-dateutil

fastapi 
uvicorn 
python-dotenv 
ollama